+++ mType/value/ValueType.hpp
@@ -47,7 +47,7 @@ namespace value
     };
 
     // Runtime value that can hold different types
-    using Value = std::variant<int64_t, float, bool, std::string, InternedString, std::monostate,
+    using Value = std::variant<int64_t, double, bool, std::string, InternedString, std::monostate,
                                std::shared_ptr<runtimeTypes::klass::ObjectInstance>,
                                std::shared_ptr<ValueObject>,
                                std::shared_ptr<NativeArray>,
